requirements.txt
langchain>=0.2.0
langchain-community>=0.0.19
openai
streamlit
pandas
python-dotenv